styles the header
