+++ src/app/components/header/header-navigation-link/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Link from "next/link";
 import { Typography } from "@mui/material";
-import { headerLinkStyles, headerLinkTypographyStyles } from "../styles";
+import { headerLinkStyles, headerLinkTypographyStyles } from "./styles";
 
 interface HeaderNavigationLinkProps {
     PATH: string;
@@ -11,11 +11,7 @@ interface HeaderNavigationLinkProps {
 const HeaderNavigationLink = ({ PATH, LABEL }: HeaderNavigationLinkProps) => {
     return (
         <Link href={PATH} passHref style={headerLinkStyles}>
-            <Typography
-                variant="body1"
-                component="a"
-                sx={headerLinkTypographyStyles}
-            >
+            <Typography variant="h4" sx={headerLinkTypographyStyles}>
                 {LABEL}
             </Typography>
         </Link>
