+++ src/app/components/header/header-navigation-link/styles.tsx
@@ -1,0 +1,11 @@
+export const headerLinkTypographyStyles = {
+    "&:hover": {
+        color: "#c6a47e",
+    },
+    fontFamily: "'Libre Baskerville', serif;",
+};
+
+export const headerLinkStyles = {
+    textDecoration: "none",
+    color: "#000",
+};
